requirements.txt
django
django-rest-framework
djangorestframework-jsonapi
jsonschema
coverage
